cmodule: Move out of GumJS and into Gum
The CModule API doesn't depend on the JS bindings, and is useful to
consumers of the C API on its own.
