+++ gum/meson.build
@@ -42,6 +42,7 @@ gum_headers = [
   'gumsysinternals.h',
   'gumthreadregistry.h',
   'gumtls.h',
+  'gumcmodule.h',
 ]
 
 gum_sources = [
@@ -95,6 +96,7 @@ gum_sources = [
   'arch-arm64' / 'gumarm64relocator.c',
   'arch-mips' / 'gummipswriter.c',
   'arch-mips' / 'gummipsrelocator.c',
+  'gumcmodule.c',
 ]
 
 gum_backend_headers = []
@@ -358,6 +360,54 @@ if libdwarf_dep.found()
   ]
 endif
 
+gum_cmodule_runtime_sources = [
+  'runtime' / 'cmodule-tcc' / 'inttypes.h',
+  'runtime' / 'cmodule-tcc' / 'stdint.h',
+  'runtime' / 'cmodule-tcc' / 'stdio.h',
+  'runtime' / 'cmodule-tcc' / 'stdlib.h',
+  'runtime' / 'cmodule-tcc' / 'string.h',
+  'runtime' / 'cmodule' / 'glib.h',
+  'runtime' / 'cmodule' / 'gum' / 'gumdefs.h',
+  'runtime' / 'cmodule' / 'gum' / 'gumprocess.h',
+  'runtime' / 'cmodule' / 'gum' / 'gummodule.h',
+  'runtime' / 'cmodule' / 'gum' / 'gummodulemap.h',
+  'runtime' / 'cmodule' / 'gum' / 'gummemory.h',
+  'runtime' / 'cmodule' / 'gum' / 'guminterceptor.h',
+  'runtime' / 'cmodule' / 'gum' / 'gumstalker.h',
+  'runtime' / 'cmodule' / 'gum' / 'gummetalarray.h',
+  'runtime' / 'cmodule' / 'gum' / 'gummetalhash.h',
+  'runtime' / 'cmodule' / 'gum' / 'gumspinlock.h',
+  'runtime' / 'cmodule' / 'gum' / 'gumtls.h',
+  'runtime' / 'cmodule' / 'json-glib' / 'json-glib.h',
+]
+
+if libtcc_dep.found()
+  if libtcc_dep.type_name() == 'internal'
+    libtcc_incdir = meson.project_source_root() / 'subprojects' / 'tinycc' / 'include'
+  else
+    libtcc_incdir = libtcc_dep.get_variable('libdir') / 'tcc' / 'include'
+  endif
+else
+  libtcc_incdir = ''
+endif
+
+gum_cmodule_runtime = custom_target('gum-cmodule-runtime',
+  input: gum_cmodule_runtime_sources,
+  output: ['gumcmodule-runtime.h'],
+  command: [
+    python,
+    files('generate-cmodule-runtime.py'),
+    meson.current_build_dir(),
+    meson.current_source_dir(),
+    meson.project_source_root() / 'gum',
+    capstone_dep.get_variable('includedir') / 'capstone',
+    libtcc_incdir,
+    host_arch,
+  ],
+)
+
+gum_sources += [gum_cmodule_runtime]
+
 install_headers(gum_headers + gum_backend_headers, subdir: install_header_subdir)
 install_headers(gum_x86_headers, subdir: install_header_subdir / 'arch-x86')
 install_headers(gum_arm_headers, subdir: install_header_subdir / 'arch-arm')
@@ -375,7 +425,11 @@ gum_enum_h = gum_enums[1]
 gum_enums_dep = declare_dependency(sources: [gum_enum_h])
 
 gum_public_deps = [glib_dep, gobject_dep, capstone_dep, gum_enums_dep]
-gum_internal_deps = [glib_dep, gobject_dep, capstone_dep, threads_dep]
+gum_internal_deps = [glib_dep, gobject_dep, gio_dep, gio_os_package_dep, json_glib_dep,
+  capstone_dep, threads_dep]
+if libtcc_dep.found()
+  gum_internal_deps += libtcc_dep
+endif
 if ffi_dep.found()
   gum_internal_deps += ffi_dep
 endif
@@ -458,7 +512,7 @@ pkg.generate(gum,
   version: frida_version,
   description: 'Low-level code instrumentation library',
   requires: ['glib-2.0', 'gobject-2.0', 'gio-2.0', 'capstone'],
-  requires_private: extra_requires_private,
+  requires_private: [gio_os_package_name, 'json-glib-1.0'] + extra_requires_private,
   subdirs: install_header_basedir,
   extra_cflags: extra_cflags,
   libraries_private: extra_libs_private,
