+++ gum/runtime/cmodule-tcc/string.h
