+++ gum/runtime/cmodule/gum/gummemory.h
