+++ gum/runtime/cmodule/gum/gummetalarray.h
