+++ gum/runtime/cmodule/gum/gummetalhash.h
