+++ gum/runtime/cmodule/gum/gumspinlock.h
